refactor: cleanup imports and components
