+++ components/__init__.py
@@ -2,6 +2,7 @@
 from PyQt6.QtCore import *
 from PyQt6.QtGui import *
 from PyQt6.QtWidgets import *
+from PyQt6.QtSvgWidgets import *
 
 # Submodules
 from . import alignment
@@ -13,7 +14,8 @@ from .application_window import ApplicationWindow
 from .container import Container
 from .icon_button import IconButton
 from .title_bar import TitleBar
-from .data_frame import DataFrame
 from .button import Button
 from .dropdown import Dropdown
 from .checkbox import CheckBox
+from .editor import Editor
+from .data_frame import DataFrame
