+++ components/data_frame.py
@@ -1,21 +1,18 @@
 from . import (
+    Component,
+    Editor,
     QTableWidget,
     QTableWidgetItem,
     QAbstractItemView,
     QStyledItemDelegate,
-    QLineEdit,
-    QIntValidator,
     Qt,
     QHeaderView,
-    QPalette,
-    QColor,
 )
+from .alignment import *
 import numpy
 import pandas
 from itertools import product
 from .settings import theme
-from .editor import Editor
-from .component import Component
 
 
 class NumericDelegate(QStyledItemDelegate):
@@ -45,7 +42,7 @@ class NumericDelegate(QStyledItemDelegate):
         text = index.model().data(index, Qt.ItemDataRole.DisplayRole)
         if text == "":
             text = self.placeholder
-        painter.drawText(option.rect, Qt.AlignmentFlag.AlignCenter, str(text))
+        painter.drawText(option.rect, AlignCenter, str(text))
         painter.restore()
 
 
@@ -58,7 +55,7 @@ class DataFrame(QTableWidget, metaclass=Component):
         radius=8,
         color=theme.text_foreground,
         selection_color=theme.context_color,
-        bg_color="transparent", #theme.bg_two,
+        bg_color="transparent",
         header_horizontal_color=theme.dark_two,
         header_vertical_color=theme.dark_two,
         bottom_line_color=theme.bg_three,
@@ -72,20 +69,19 @@ class DataFrame(QTableWidget, metaclass=Component):
         self.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Stretch)
         self.horizontalHeader().setSectionsClickable(False)
         self.horizontalHeader().setMinimumHeight(30)
-        
+
         # Vertical header (index) settings
         if not flexible_rows:
             self.verticalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Fixed)
         self.verticalHeader().setSectionsClickable(False)
         self.verticalHeader().setMinimumSectionSize(30)
-        
+
         # Window settings
         self.setSelectionMode(QAbstractItemView.SelectionMode.NoSelection)
         self.setItemDelegate(NumericDelegate(self, maxlen=10, placeholder=placeholder))
         self.index = []
         self.label = []
         self.data = data
-
 
     @property
     def data(self):
